+++ src/Events/SMSSendingFailedEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace NotificationChannels\Cmsms\Events;
+
+use Illuminate\Foundation\Events\Dispatchable;
+
+class SMSSendingFailedEvent
+{
+    use Dispatchable;
+
+    public function __construct(public string $response)
+    {
+    }
+}