+++ tests/CmsmsMessageTest.php
@@ -2,7 +2,6 @@
 
 namespace NotificationChannels\Cmsms\Test;
 
-use Illuminate\Support\Arr;
 use NotificationChannels\Cmsms\CmsmsMessage;
 use NotificationChannels\Cmsms\Exceptions\InvalidMessage;
 use PHPUnit\Framework\TestCase;
@@ -22,7 +21,7 @@ class CmsmsMessageTest extends TestCase
     {
         $message = CmsmsMessage::create('Foo');
 
-        $this->assertEquals('Foo', Arr::get($message->toXmlArray(), 'BODY'));
+        $this->assertEquals('Foo', $message->getBody());
     }
 
     /** @test */
@@ -31,7 +30,7 @@ class CmsmsMessageTest extends TestCase
         $message = CmsmsMessage::create('Foo');
 
         $this->assertInstanceOf(CmsmsMessage::class, $message);
-        $this->assertEquals('Foo', Arr::get($message->toXmlArray(), 'BODY'));
+        $this->assertEquals('Foo', $message->getBody());
     }
 
     /** @test */
@@ -39,7 +38,7 @@ class CmsmsMessageTest extends TestCase
     {
         $message = CmsmsMessage::create('Bar');
 
-        $this->assertEquals('Bar', Arr::get($message->toXmlArray(), 'BODY'));
+        $this->assertEquals('Bar', $message->getBody());
     }
 
     /** @test */
@@ -47,7 +46,7 @@ class CmsmsMessageTest extends TestCase
     {
         $message = CmsmsMessage::create()->originator('APPNAME');
 
-        $this->assertEquals('APPNAME', Arr::get($message->toXmlArray(), 'FROM'));
+        $this->assertEquals('APPNAME', $message->getOriginator());
     }
 
     /** @test */
@@ -71,7 +70,7 @@ class CmsmsMessageTest extends TestCase
     {
         $message = CmsmsMessage::create()->reference('REFERENCE123');
 
-        $this->assertEquals('REFERENCE123', Arr::get($message->toXmlArray(), 'REFERENCE'));
+        $this->assertEquals('REFERENCE123', $message->getReference());
     }
 
     /** @test */
@@ -98,28 +97,12 @@ class CmsmsMessageTest extends TestCase
         CmsmsMessage::create()->reference('@#$*A*Sjks87');
     }
 
-    /** @test */
-    public function it_can_set_tariff()
-    {
-        $message = CmsmsMessage::create()->tariff(12);
-
-        $this->assertEquals(12, $message->getTariff());
-    }
-
-    /** @test */
-    public function it_can_set_an_empty_tariff()
-    {
-        $message = CmsmsMessage::create()->tariff(0);
-
-        $this->assertEquals(0, $message->getTariff());
-    }
-
     public function it_can_set_multipart()
     {
         $message = CmsmsMessage::create()->multipart(1, 4);
 
-        $this->assertEquals(1, Arr::get($message->toXmlArray(), 'MINIMUMNUMBEROFMESSAGEPARTS'));
-        $this->assertEquals(4, Arr::get($message->toXmlArray(), 'MAXIMUMNUMBEROFMESSAGEPARTS'));
+        $this->assertEquals(1, $message->getMinimumNumberOfMessageParts());
+        $this->assertEquals(4, $message->getMaximumNumberOfMessageParts());
     }
 
     public function it_cannot_set_more_than_8_parts_to_multipart()
@@ -135,15 +118,5 @@ class CmsmsMessageTest extends TestCase
         $this->expectException(InvalidMessage::class);
 
         CmsmsMessage::create()->multipart(4, 3);
-    }
-
-    /** @test */
-    public function it_xml_contains_only_filled_parameters()
-    {
-        $message = CmsmsMessage::create('Foo');
-
-        $this->assertEquals([
-            'BODY' => 'Foo',
-        ], $message->toXmlArray());
     }
 }